# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MoneyKeeper API",
    "description": "This is a simple MoneyKeeper software.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://alnat.dev"
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/user/": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Получение списка пользователей",
        "operationId": "getUserList",
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "put": {
        "tags": [
          "User API"
        ],
        "summary": "Обновление пользователя",
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Пользователь успешно обновлен"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "post": {
        "tags": [
          "User API"
        ],
        "summary": "Создание нового пользователя",
        "operationId": "addUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Пользователь успешно создан"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/user/search": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Получение пользователя по имени",
        "operationId": "getUserByName",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "Имя пользователя",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "admin"
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Пользователя с таким именем не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/user/{userID}": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Получение пользователя по идентификатору",
        "operationId": "getUserByID",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "Идентификатор пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Пользователь с таким идентификатором не найден",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "delete": {
        "tags": [
          "User API"
        ],
        "summary": "Удаление пользователя",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "Идентификатор пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не корректен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Пользователь с таким идентификатором не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Удаление успешно выполнено"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/user/group/": {
      "post": {
        "tags": [
          "User API"
        ],
        "summary": "Добавление пользователя в группу",
        "operationId": "addUserToGroup",
        "parameters": [
          {
            "name": "userID",
            "in": "query",
            "description": "Идентификатор пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          },
          {
            "name": "userGroupID",
            "in": "query",
            "description": "Идентификатор группы",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не корректен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Пользователь с таким идентификатором не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Добавление успешно выполнено"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "delete": {
        "tags": [
          "User API"
        ],
        "summary": "Удаление пользователя из группы",
        "operationId": "removeUserFromGroup",
        "parameters": [
          {
            "name": "userID",
            "in": "query",
            "description": "Идентификатор пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          },
          {
            "name": "userGroupID",
            "in": "query",
            "description": "Идентификатор группы",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не корректен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Пользователь с таким идентификатором не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Удаление успешно выполнено"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/usergroup/search": {
      "get": {
        "tags": [
          "User Group API"
        ],
        "summary": "Получение группы по идентификатору",
        "operationId": "getUserGroupByKey",
        "parameters": [
          {
            "name": "userGroupKey",
            "in": "query",
            "description": "Идентификатор группы",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "admin"
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Группы с таким именем не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroup"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroup"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/usergroup/{userGroupID}": {
      "get": {
        "tags": [
          "User Group API"
        ],
        "summary": "Получение группы по идентификатору",
        "operationId": "getUserGroupByID",
        "parameters": [
          {
            "name": "userGroupID",
            "in": "path",
            "description": "Идентификатор группы",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Группы с таким идентификатором не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroup"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserGroup"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "delete": {
        "tags": [
          "User Group API"
        ],
        "summary": "Удаление группы",
        "operationId": "deleteUserGroup",
        "parameters": [
          {
            "name": "userGroupID",
            "in": "path",
            "description": "Идентификатор группы",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не корректен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Группы с таким идентификатором не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Удаление успешно выполнено"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/usergroup/": {
      "get": {
        "tags": [
          "User Group API"
        ],
        "summary": "Получение списка пользовательских групп",
        "operationId": "getUserGroupList",
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserGroup"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserGroup"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "put": {
        "tags": [
          "User Group API"
        ],
        "summary": "Обновление группы",
        "operationId": "updateUserGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Группа успешно обновлена"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "post": {
        "tags": [
          "User Group API"
        ],
        "summary": "Создание новой группы",
        "operationId": "addUserGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Группа успешно создана"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки"
          }
        }
      },
...
      "User": {
        "required": [
          "enabled",
          "password",
          "userGroupList",
          "userID",
          "username"
        ],
        "type": "object",
        "properties": {
          "userID": {
            "type": "integer",
            "description": "Идентификатор пользователя",
            "format": "int32",
            "example": 1
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "1"
          },
          "username": {
            "pattern": "^[А-Яа-яЁёA-Za-z0-9 -]{0,255}$",
            "type": "string",
            "description": "Логин пользователя",
            "example": "administrator"
          },
          "enabled": {
            "type": "boolean",
            "description": "Флаг активности пользователя"
          },
          "userGroupList": {
            "type": "array",
            "description": "Список групп, в которые входит пользователь",
            "items": {
              "$ref": "#/components/schemas/UserGroup"
            }
          }
        },
        "description": "Измененный пользователь"
      },
      "UserGroup": {
        "required": [
          "key",
          "name",
          "userGroupID",
          "userOperationList"
        ],
        "type": "object",
        "properties": {
          "userGroupID": {
            "type": "integer",
            "description": "Идентификатор группы",
            "format": "int32",
            "example": 1
          },
          "key": {
            "type": "string",
            "description": "Идентификатор группы",
            "example": "user"
          },
          "name": {
            "type": "string",
            "description": "Имя группы",
            "example": "Пользователи"
          },
          "userOperationList": {
            "type": "array",
            "description": "Список доступных операций для этой группы",
            "items": {
              "type": "string",
              "description": "Список доступных операций для этой группы",
              "enum": [
                "DASHBOARD",
                "TRANSACTION_LIST",
                "TRANSACTION",
                "TRANSACTION_DELETE",
                "TRANSACTION_CHANGE",
                "TRANSACTION_CREATE",
                "CATEGORY",
                "CATEGORY_DELETE",
                "CATEGORY_CHANGE",
                "CATEGORY_CREATE",
                "ACCOUNT_LIST",
                "ACCOUNT",
                "ACCOUNT_DELETE",
                "ACCOUNT_CHANGE",
                "ACCOUNT_CREATE",
                "USER_LIST",
                "USER",
                "USER_DELETE",
                "USER_CHANGE",
                "USER_CREATE",
                "USER_GROUP_LIST",
                "USER_GROUP",
                "USER_GROUP_DELETE",
                "USER_GROUP_CHANGE",
                "USER_GROUP_CREATE"
              ]
            }
          }
        },
        "description": "Новая группа"
      }
    }
  }
}